[TTS] Add license to ported EnCodec code (#7197)
